Split up tests and rely less on http
